Somewhat reasonable

+++ src/examples/n_gram.rs
@@ -217,7 +217,7 @@ where
 
             if sleep_time > 0 && epoch != 0 {
                 println!("Taking a nap");
-                thread::sleep(Duration::from_secs(sleep_time as u64));
+                thread::sleep(Duration::from_secs(sleep_time));
                 println!("Awake again");
             }
         };
